iframebreaker/requirements.txt
minio==7.2.15
pika==1.3.2
python-dotenv==1.1.0
pathvalidate==3.3.1
